Added SubscriptionPolicy class. Changed to ServiceController.ExecuteAsync
The SubscriptionPolicy encapsulates the logic for dealing with a
subscription being dropped. In some cases we want to be able to retry the
subscription to a maximum specified number of retries.

To make the Plugin as asynchronous as possible I have switched from using
ServiceController.Execute to ServiceController.ExecuteAsync. A
BasicRequest:IRequest is passed as recommended by Demis:
https://forums.servicestack.net/t/using-servicestack-architecture-to-route-events-instead-of-using-reflection/2185/7