Added ability to populate a read model from events. Set access modifiers to internal where appropriate.
By (1) specifying a ReadModelSubscription, (2) creating a class that
inherits from ServiceStack.Service with DTO types that map to the desired
events, (3) creating a viewmodel to be persisted, and (4) declaring an
IProjectionWriter<TId, TViewModel> a read model can be populated.

Also refactored classes to use the 'internal' access modifier if they
should not be accessible from outside the plugin. Also added code
comments and attributions.

EventTypes is now a static class and allows for scanning types that
implement IAggregateEvent.