Removed access to underlying EventStoreConnection.
A reference to EventStoreConnection from the EventStore.Client.API was
exposed via IEventStoreRepository which meant the consumer needed to
deal with StreamSlices etc instead of getting a stream of events.

I've therefore added two new methods: ReadSliceFromStreamAsync and
DeleteStreamAsync to provide the functionality that was being used by
the integration tests.