Removed references to embedded client in Feature
When I refactored the integration tests to run against an embedded
instance of EventStore instead of a live instance I added a reference to
the embedded client to the EventStoreFeature and passed in a builder
node into the Feature where it was used to create an instance of
an EmbeddedEventStoreConnection.

I wasn't ideal to have the plugin knowing about something that is only
required when running tests so I have removed the references to the
embedded client from the plugin and have now refactored the
EventStoreFeature to accept an object that implements
IEventStoreConnection. This means that an EmbeddedEventStoreConnection
can be passed in from the integration tests project.